requirements.txt
numpy
pillow
paddlepaddle
paddleocr
flask
gevent
